Working call to bitpos api
