+++ src/Win8POS/POS8/POS8.csproj
@@ -102,15 +102,14 @@
     <Prefer32Bit>true</Prefer32Bit>
   </PropertyGroup>
   <ItemGroup>
-    <!-- A reference to the entire .Net Framework and Windows SDK are automatically included -->
-  </ItemGroup>
-  <ItemGroup>
     <Compile Include="App.xaml.cs">
       <DependentUpon>App.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Models\BitPOS\Order.cs" />
     <Compile Include="Models\ISettings.cs" />
-    <Compile Include="Models\Order.cs" />
+    <Compile Include="ViewModels\Order.cs" />
     <Compile Include="Models\BitPOSSettings.cs" />
+    <Compile Include="Repositories\BitPOS.cs" />
     <Compile Include="Repositories\LocalDB.cs" />
     <Compile Include="SplitPage.xaml.cs">
       <DependentUpon>SplitPage.xaml</DependentUpon>
@@ -128,6 +127,7 @@
       <SubType>Designer</SubType>
     </AppxManifest>
     <None Include="Common\ReadMe.txt" />
+    <None Include="packages.config" />
     <None Include="POS8_TemporaryKey.pfx" />
     <Content Include="Repositories\SampleData.json" />
   </ItemGroup>
@@ -149,6 +149,11 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
+  </ItemGroup>
+  <ItemGroup>
+    <Reference Include="Newtonsoft.Json">
+      <HintPath>..\packages\Newtonsoft.Json.6.0.8\lib\netcore45\Newtonsoft.Json.dll</HintPath>
+    </Reference>
   </ItemGroup>
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '12.0' ">
     <VisualStudioVersion>12.0</VisualStudioVersion>
