+++ src/Win8POS/POS8/SplitPage.xaml
@@ -17,7 +17,15 @@
             x:Name="itemsViewSource"
             Source="{Binding Orders}"/>
     </Page.Resources>
-
+    <Page.TopAppBar>
+        <CommandBar IsSticky="True">
+            <AppBarButton Label="Refresh" Icon="Refresh" Click="AppBarButton_Click"/>
+            <AppBarButton Label="Help" Icon="Help" Click="AppBarButton_Click"/>
+            <CommandBar.SecondaryCommands>
+                <AppBarButton Label="Add" Icon="Add" Click="AppBarButton_Click"/>
+            </CommandBar.SecondaryCommands>
+        </CommandBar>
+    </Page.TopAppBar>
     <!--
         This grid acts as a root panel for the page that defines two rows:
         * Row 0 contains the back button and page title
@@ -51,6 +59,12 @@
                         AutomationProperties.Name="Back"
                         AutomationProperties.AutomationId="BackButton"
                         AutomationProperties.ItemType="Navigation Button"/>
+            <Button x:Name="addButton" Margin="39,59,39,0" Command="{Binding NavigationHelper.GoBackCommand, ElementName=pageRoot}"
+                        Style="{StaticResource NavigationBackButtonNormalStyle}"
+                        VerticalAlignment="Top"
+                        AutomationProperties.Name="Add"
+                        AutomationProperties.AutomationId="BackButton"
+                        AutomationProperties.ItemType="Navigation Button"/>
             <TextBlock x:Name="pageTitle" Text="{Binding Title}" Style="{StaticResource HeaderTextBlockStyle}" Grid.Column="1" 
                         IsHitTestVisible="false" TextWrapping="NoWrap" VerticalAlignment="Bottom" Margin="0,0,0,40"/>
         </Grid>
@@ -74,11 +88,12 @@
                             <ColumnDefinition Width="Auto"/>
                             <ColumnDefinition Width="*"/>
                         </Grid.ColumnDefinitions>
-                        <Border Background="{ThemeResource ListViewItemPlaceholderBackgroundThemeBrush}" Width="60" Height="60">
+                        <Border Background="{ThemeResource ListViewItemPlaceholderBackgroundThemeBrush}" Width="120" Height="60">
                             <TextBlock Text="{Binding Amount}"></TextBlock>
                         </Border>
                         <StackPanel Grid.Column="1" Margin="10,0,0,0">
-                            <TextBlock Text="{Binding Title}" Style="{StaticResource TitleTextBlockStyle}" TextWrapping="NoWrap" MaxHeight="40"/>
+                            <TextBlock Text="{Binding Reference}" Style="{StaticResource TitleTextBlockStyle}" TextWrapping="NoWrap" MaxHeight="40"/>
+                            <TextBlock Text="bla"></TextBlock>
                         </StackPanel>
                     </Grid>
                 </DataTemplate>
@@ -89,6 +104,7 @@
                 </Style>
             </ListView.ItemContainerStyle>
         </ListView>
+
 
         <!-- Details for selected item -->
         <!--<ScrollViewer
