+++ src/Win8POS/POS8/bin/Debug/AppX/vs.appxrecipe
@@ -17,13 +17,17 @@
     <AppXManifest Include="C:\Users\Lucas Cullen\Source\GitHub\api-doc\src\Win8POS\POS8\bin\Debug\AppxManifest.xml">
       <PackagePath>AppxManifest.xml</PackagePath>
       <ReRegisterAppIfChanged>true</ReRegisterAppIfChanged>
-      <Modified>2015-02-26T02:50:21.285</Modified>
+      <Modified>2015-02-27T11:21:47.767</Modified>
     </AppXManifest>
   </ItemGroup>
   <ItemGroup>
     <AppxPackagedFile Include="C:\Users\Lucas Cullen\Source\GitHub\api-doc\src\Win8POS\POS8\bin\Debug\POS8.exe">
       <PackagePath>POS8.exe</PackagePath>
-      <Modified>2015-02-26T03:07:20.672</Modified>
+      <Modified>2015-03-01T06:02:00.355</Modified>
+    </AppxPackagedFile>
+    <AppxPackagedFile Include="C:\Users\Lucas Cullen\Source\GitHub\api-doc\src\Win8POS\packages\Newtonsoft.Json.6.0.8\lib\netcore45\Newtonsoft.Json.dll">
+      <PackagePath>Newtonsoft.Json.dll</PackagePath>
+      <Modified>2015-02-27T08:49:06.189</Modified>
     </AppxPackagedFile>
     <AppxPackagedFile Include="C:\Users\Lucas Cullen\Source\GitHub\api-doc\src\Win8POS\POS8\Repositories\SampleData.json">
       <PackagePath>Repositories\SampleData.json</PackagePath>
@@ -63,19 +67,15 @@
     </AppxPackagedFile>
     <AppxPackagedFile Include="C:\Users\Lucas Cullen\Source\GitHub\api-doc\src\Win8POS\POS8\bin\Debug\App.xbf">
       <PackagePath>App.xbf</PackagePath>
-      <Modified>2015-02-26T02:45:52.981</Modified>
-    </AppxPackagedFile>
-    <AppxPackagedFile Include="C:\Users\Lucas Cullen\Source\GitHub\api-doc\src\Win8POS\POS8\bin\Debug\ItemsPage.xbf">
-      <PackagePath>ItemsPage.xbf</PackagePath>
-      <Modified>2015-02-26T02:45:52.984</Modified>
+      <Modified>2015-02-27T11:21:47.264</Modified>
     </AppxPackagedFile>
     <AppxPackagedFile Include="C:\Users\Lucas Cullen\Source\GitHub\api-doc\src\Win8POS\POS8\bin\Debug\SplitPage.xbf">
       <PackagePath>SplitPage.xbf</PackagePath>
-      <Modified>2015-02-26T03:01:02.170</Modified>
+      <Modified>2015-03-01T05:55:22.738</Modified>
     </AppxPackagedFile>
     <AppxPackagedFile Include="C:\Users\Lucas Cullen\Source\GitHub\api-doc\src\Win8POS\POS8\bin\Debug\resources.pri">
       <PackagePath>resources.pri</PackagePath>
-      <Modified>2015-02-26T02:50:21.235</Modified>
+      <Modified>2015-02-27T11:21:47.703</Modified>
     </AppxPackagedFile>
   </ItemGroup>
   <ItemGroup>
